Transformer
This project implements an image captioning system that generates natural language descriptions from input images. A ResNet50 convolutional neural network is used to extract high-level visual features from the images, which are then passed to a Transformer-based decoder to produce captions word by word. The model is trained on the Flickr dataset with data augmentation and optimized using cross-entropy loss. A custom vocabulary builder is used to tokenize and numericalize the captions, and BLEU score is employed for evaluation.

The aim of this project is to bridge computer vision and natural language processing using modern deep learning architectures.